Move Google sign in to NavBar
