+++ src/components/NavBar.js
@@ -1,6 +1,7 @@
-import React, { useState } from 'react';
+import React from 'react';
 import 'bootstrap/dist/css/bootstrap.css';
 import AddClassForm from './AddClassForm.js'
+import { gapi } from 'gapi-script';
 
 import {
   Navbar,
@@ -13,16 +14,85 @@ import {
   DropdownMenu,
 } from 'reactstrap';
 
-const NavBar = (props) => {
-  const [isOpen, setIsOpen] = useState(false);
+class NavBar extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      isOpen: false,
+      isSignedIn: false,
+    }
+  }
 
-  const toggle = () => setIsOpen(!isOpen);
+  componentDidMount() {
 
+    const successCallback = this.onSuccess.bind(this);
+    
+    window.gapi.load('auth2', () => {
+      this.auth2 = gapi.auth2.init({
+        client_id: '724928841047-2a73bpb9m7hat1v4rhdrkefrtaedmp26.apps.googleusercontent.com',
+      })
+
+      // this.auth2.attachClickHandler(document.querySelector('#loginButton'), {}, this.onLoginSuccessful.bind(this))
+
+      this.auth2.then(() => {
+        console.log('on init');
+        this.setState({
+          isSignedIn: this.auth2.isSignedIn.get(),
+        });
+      });
+    });    
+
+    window.gapi.load('signin2', function() {
+      // Method 3: render a sign in button
+      // using this method will show Signed In if the user is already signed in
+      var opts = {
+        width: 200,
+        height: 50,
+        client_id: 'YOUR_CLIENT_ID.apps.googleusercontent.com',
+        onsuccess: successCallback
+      }
+      gapi.signin2.render('loginButton', opts)
+    })
+  }
+
+  onSuccess() {
+    console.log('on success')
+    this.setState({
+      isSignedIn: true,
+      err: null
+    })
+  }
+
+  onLoginFailed(err) {
+    this.setState({
+      isSignedIn: false,
+      error: err,
+    })
+  }
+
+  getContent() {
+    if (this.state.isSignedIn) {
+      return <p>hello user, you're signed in </p>
+    } else {
+      return (
+        <div>
+          <p>You are not signed in. Click here to sign in.</p>
+          <button id="loginButton">Login with Google</button>
+        </div>
+      )
+    }
+  }
+
+  toggle() {
+    this.setState({isOpen: !this.state.isOpen});
+  }
+
+  render() {
   return (
     <div>
       <Navbar color="light" light expand="md">
         <NavbarBrand href="/">bunkalog</NavbarBrand>
-        <NavbarToggler onClick={toggle} />
+        <NavbarToggler onClick={this.toggle} />
         <Nav>
           <NavItem>
 						<UncontrolledDropdown nav inNavbar>
@@ -34,10 +104,14 @@ const NavBar = (props) => {
               </DropdownMenu>
             </UncontrolledDropdown>
           </NavItem>
+          <NavItem>
+            {this.getContent()}
+          </NavItem>
         </Nav>
       </Navbar>
     </div>
   );
+  }
 }
 
 export default NavBar;
